+++ src/entities/user/api/userApi.ts
@@ -1,0 +1,7 @@
+import { IUser } from "../model/types"
+
+export const fetchUserInfo = async (userId: number) => {
+  const res = await fetch(`/api/users/${userId}`)
+  if (!res.ok) throw new Error("Failed to fetch user")
+  return res.json() as Promise<IUser>
+}
