+++ src/features/postManager/hook/useFilteredPosts.ts
@@ -1,0 +1,6 @@
+// features/post/api/postManagerApi.ts
+
+export { usePosts } from "@/entities/post/hooks/usePosts"
+export { useSearchPosts } from "@/entities/post/hooks/useSearchPosts"
+export { usePostByTag } from "@/entities/post/hooks/usePostByTag"
+export { usePostTags } from "@/entities/post/hooks/usePostTags"
